+++ ERP.Data/ModelsERP/ModelView/Statistics/statisticsorderviewmodel.cs
@@ -12,7 +12,7 @@ namespace ERP.Data.ModelsERP.ModelView.Statistics
         public byte? cuo_status { get; set; }
         public string cuo_status_name { get; set; }
         public DateTime? cuo_date { get; set; }
-        public float? op_total_value { get; set; }
+        public double? op_total_value { get; set; }
         public int pu_id { get; set; }
         public string pu_name { get; set; }
